+++ main.py
@@ -1,11 +1,26 @@
 import csv
 
 f = open('data.csv', 'r')
-b = []
+
+array = []
+pointer = 0
+
 with f:
     reader = csv.reader(f)
     for row in reader:
         for e in row:
-            print(e)
-            e = b.append
+          array[pointer] = e
+          pointer += 1
 
+def bubbleSort(array):
+        length = len(array)-1
+        indexminus = 0;
+        
+        for i in range(length):
+            for j in range(0, length-indexminus):
+                if array[j] > array[j+1]:
+                    temp = array[j]
+                    array[j] = array[j+1]
+                    array[j+1] = temp
+
+bubbleSort(array)
